requirements.txt
numpy
pandas
sqlalchemy
python-dotenv
pymysql
